final commit

+++ src/pages/Dashboard/StudentSection/MySelectedClasses/MySelectedClasses.jsx
@@ -2,6 +2,7 @@ import { useQuery } from "@tanstack/react-query";
 import useAuth from "../../../../Hooks/useAuth";
 import useAxiosSecure from "../../../../Hooks/useAxiosSecure";
 import SelectedClassList from "./SelectedClassList";
+import Swal from "sweetalert2";
 
 
 const MySelectedClasses = () => {
@@ -17,18 +18,37 @@ const MySelectedClasses = () => {
     })
 
 
-    
+
     const handleDelete = _id => {
 
-        fetch(`http://localhost:5000/selectedClasses/${_id}`, {
-            method: 'DELETE'
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`http://localhost:5000/selectedClasses/${_id}`, {
+                    method: 'DELETE'
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        if (data.deletedCount > 0) {
+                            Swal.fire(
+
+                                'Deleted!',
+                                'Your selected class has been deleted.',
+                                'success'
+                            )
+                            refetch()
+                        }
+                    })
+            }
         })
-            .then(res => res.json())
-            .then(data => {
-                if (data.deletedCount > 0) {
-                    refetch()
-                }
-            })
+
     }
